requirements.txt
pillow==12.0.0
pypdf==6.4.0
